feat: support showQrCode with WalletConnect
