+++ packages/web3/src/connect-button/index.zh-CN.md
@@ -10,3 +10,15 @@ group: 组件
 ## 基本使用
 
 <code src="./demos/simple.tsx"></code>
+
+## 和 wagmi 一起使用
+
+<code src="./demos/wagmi.tsx"></code>
+
+## 使用 web3modal 连接 WallectConnect
+
+<code src="./demos/web3modal.tsx"></code>
+
+## 和 @ant-design/web3-ethereum 一起使用
+
+<code src="./demos/ethereum.tsx"></code>
